entropy analysis moved to the src, as well as FixPA
