+++ examples/collapse_symplectic.jl
@@ -10,17 +10,15 @@ Simulation of a water column collapsing under its own weight onto dry bottom.
 This is, where SPH is more useful than typical mesh-based methods
 =#
 
-module symplectic
+module collapse_symplectic
 
 using Printf
 include("../src/SPHLib.jl")
 using .SPHLib
 using Parameters
 using Plots
-using Ipopt
-using JuMP
-include("FixPA.jl")
-#using ReadVTK 
+
+#using ReadVTK  #not implemented
 #using VTKDataIO
 
 #=
@@ -51,7 +49,7 @@ const eps = 1e-16
 
 ##temporal
 const dt = 0.1*h/c
-const t_end = 1.00
+const t_end = 1.0
 const dt_frame = t_end/100
 
 ##particle types
@@ -126,13 +124,13 @@ end
 
 function move!(p::Particle)
 	if p.type == FLUID
-		p.x = FixPA.rev_add(p.x, dt*p.v)
+		p.x = rev_add(p.x, dt*p.v)
 	end
 end
 
 function accelerate!(p::Particle)
 	if p.type == FLUID
-		p.v = FixPA.rev_add(p.v, 0.5*dt*(p.a + g))
+		p.v = rev_add(p.v, 0.5*dt*(p.a + g))
 	end
 end
 
@@ -185,7 +183,7 @@ function save_results!(out::SPHLib.DataStorage, sys::ParticleSystem, k::Int64)
     end
 end
 
-function main(;revert = true)
+function main(;revert = true) #if revert=true, velocities are inverted at the end of the simulation and the simulation then goes backward
 	sys = make_system()
 	out = new_pvd_file("results/collapse_fixpa")
     #initialization
@@ -214,11 +212,12 @@ function main(;revert = true)
 			push!(Ss, S)
 			push!(Ekin, energy_kinetic(sys))
 			@show(S)
+        	println()
 		end
 	end
 
 	# Plotting the velocity distribution in comparison with Maxwell-Boltzmann
-	T = plot_velocity_distr(sys, "energy_distribution_middle.pdf")
+	T = plot_velocity_distr(sys, m, "energy_distribution_middle.pdf")
 
 	# Plotting the entropy in time
 	Sred_eq_E = [(1+log(Ekin[k]/(m*length(sys.particles)))) for k in 1:length(Ss)]
@@ -243,9 +242,10 @@ function main(;revert = true)
 				S = entropy_2D_MB(distr)
 				push!(Ss_rev, S)
 				@show(S)
+				println()
 			end
 		end
-		plot_velocity_distr(sys, "energy_distribution_final.pdf")
+		plot_velocity_distr(sys, m, "energy_distribution_final.pdf")
 
 		# Plotting the entropy in time
 		p = plot(times, [Ss Ss_rev Sred_eq_T Sred_eq_E], label = ["entropy forward" "entropy backward" "S_eq(T)" "S_eq(E)"], legend=:bottomright)
@@ -255,137 +255,6 @@ function main(;revert = true)
 	save_pvd_file(out)
 
 end ## function main
-
-#function boltzmann(beta, e)::Float64
-#	return beta*exp(-e*beta)
-#end
-
-"""
-	Histogram(xs::Vector{Float64}, ys::Vector{Float64},N,dx)
-
-Histogram structure storing ``N`` x values ``xs`` with uniform bin width ``dx`` and ``N`` y values ``ys``.
-"""
-struct Histogram
-	xs::Vector{Float64}
-	ys::Vector{Float64}
-	N::Int64
-	dx::Float64
-end
-
-"""
-	velocity_histogram(sys::ParticleSystem; v_max = 0, N = 10)
-
-Building the histrogram of 2D velocities (norms) with ``v_max`` the maximum velocity in the histogram and ``N`` bins.
-"""
-function velocity_histogram(sys::ParticleSystem; v_max = 0.0, N = 100)::Histogram
-	if v_max == 0.0 # if v_max = 0, find the maximum velocity of the particles
-		for k in 1:length(sys.particles) 
-			v = norm(sys.particles[k].v)
-			if v > v_max
-				v_max = v
-			end
-		end
-	end
-
-	# Find the heights of the histogram bins
-	dv = v_max/N # velocity increment between the bins
-	vs = 0.:dv:v_max
-	ns = zeros(length(vs))
-	for k in 1:length(sys.particles)
-		v = norm(sys.particles[k].v)
-		n = Int64(round(v/dv))
-		if 1 <= n <= length(ns)
-			ns[n] += 1.0/(dv*length(sys.particles))
-		end
-	end
-
-	return Histogram(vs,ns,100,dv)
-end
-
-
-
-"""
-	kB
-
-Boltzmann constant (in the SI units)
-"""
-const kB = 1.380649e-23
-
-"""
-	entropy(fMB::Histogram)::Float64
-
-Calculate Boltzmann entropy of a 2D Maxwell-Boltzmann distribution approximated by an ``fMB`` histogram.
-"""
-function entropy_2D_MB(fMB::Histogram)::Float64
-	@assert(fMB.xs[1] == 0) # Assuming that the histogram starts at zero velocity
-
-	S = 0.0
-
-	# Approximating the reduced entropy near v=0, where a numerical singularity could appear
-	fMBder = (fMB.ys[2]-fMB.ys[1])/fMB.dx
-	if fMBder > 0
-		S = - fMB.ys[1] * (log(fMBder)*fMB.dx - fMBder*(fMB.dx^3)/6)
-	end
-
-	# Approximating the rest of entropy
-	for k in 2:length(fMB.xs)
-		if fMB.xs[k] != 0
-			if fMB.ys[k] > 0
-				S += -fMB.ys[k] * log(fMB.ys[k]/fMB.xs[k]) * fMB.dx
-			end
-		end
-	end
-
-	return S
-end
-
-""" 
-	plot_velocity_distr(sys::ParticleSystem, name::String)
-
-Plots the distribution of velocity magnitudes among the particles of ``sys`` and saves the resulting pdf,
-together with a fit of the Maxwell-Boltzmann distribution, to file ``name``.
-Returns the temperature.
-"""
-function plot_velocity_distr(sys::ParticleSystem, name::String; v_max = 0.0)::Float64
-	distr = velocity_histogram(sys, v_max=v_max, N = 100)
-	
-	# fitting the histogram to a 2D Maxwell-Boltzmann distribution
-	model = Model(Ipopt.Optimizer)
-	@variable(model, beta)
-	@NLobjective(
-			model,
-			Min,
-			sum((distr.ys[i] - m*beta*distr.xs[i]*exp(-0.5*m*beta*distr.xs[i]^2))^2 for i in 1:length(distr.xs)),
-		) 
-	optimize!(model)
-	beta = value(beta)
-
-	# Plotting both the actual histogram and the fitted Maxwell-Boltzmann distribution
-	ns_boltz = zeros(length(distr.xs))
-	for i in 1:length(ns_boltz)
-		ns_boltz[i] = m*beta*distr.xs[i]*exp(-0.5*m*beta*distr.xs[i]^2)
-	end
-	@show(beta)
-	T = 1/(beta*kB)
-	@show(T)
-
-	p = plot(distr.xs, [distr.ys ns_boltz], label = ["data" "Maxwell-Boltzmann, T="*string(T)])
-	savefig(p, name)
-	return T
-end
-
-"""
-	plot_velocity_distr(path::String, name:: String)
-
-Plot velocity distribution of particles loaded from a VTK file.
-"""
-function plot_velocity_distr(path::String, name::String)
-	@error "Not working, sorry"
-	domain = Rectangle(-box_width, -box_width, 2*box_width, 3*box_height) 
-	sys = ParticleSystem(Particle, domain, h)
-	read_vtk!(sys, path, x -> Particle(x = x, type = 0.0))
-	plot_velocity_distr(sys, name)  #not yet working
-end
 
 end ## module
